+++ Bouhm.Shared/Integrations/GenericModConfigMenu/IGenericModConfigMenuApi.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using StardewModdingAPI;
+using StardewModdingAPI.Utilities;
 
 namespace Bouhm.Shared.Integrations.GenericModConfigMenu;
 
@@ -46,6 +47,15 @@ public interface IGenericModConfigMenuApi
     /// <param name="tooltip">The tooltip text shown when the cursor hovers on the field, or <c>null</c> to disable the tooltip.</param>
     /// <param name="fieldId">The unique field ID for use with <c>OnFieldChanged</c>, or <c>null</c> to auto-generate a randomized ID.</param>
     void AddKeybind(IManifest mod, Func<SButton> getValue, Action<SButton> setValue, Func<string> name, Func<string> tooltip = null, string fieldId = null);
+
+    /// <summary>Add a key binding list at the current position in the form.</summary>
+    /// <param name="mod">The mod's manifest.</param>
+    /// <param name="getValue">Get the current value from the mod config.</param>
+    /// <param name="setValue">Set a new value in the mod config.</param>
+    /// <param name="name">The label text to show in the form.</param>
+    /// <param name="tooltip">The tooltip text shown when the cursor hovers on the field, or <c>null</c> to disable the tooltip.</param>
+    /// <param name="fieldId">The unique field ID for use with <c>OnFieldChanged</c>, or <c>null</c> to auto-generate a randomized ID.</param>
+    void AddKeybindList(IManifest mod, Func<KeybindList> getValue, Action<KeybindList> setValue, Func<string> name, Func<string>? tooltip = null, string? fieldId = null);
 
     /// <summary>Add an integer option at the current position in the form.</summary>
     /// <param name="mod">The mod's manifest.</param>